standardize on director terminology
there are member-based nonprofits, and memberless nonprofits. We are memberless. Additionally, some nonprofits call their people on the Board "members". This is confusing.

it's too mentally taxing to have both a "Board of Directors" and "Board members". All the people on the Board are directors, so let's just standardize on calling them directors. BoDs and BDs. Not BoDs and BMs.